# refactor(dht): Optimize storer analysis